air-quality-pipeline-bbsr/requirements.txt
fastapi
uvicorn[standard]
asyncpg
kafka-python
pyflink
